Direct2D sample now compiles perfectly on FPC

+++ Samples_FPC/Direct2D/TextAnimationSample/RingBuffer.pas
@@ -81,7 +81,7 @@ end;
 
 function TRingBuffer.GetLast: LONGLONG;
 begin
-    Result := FElements[(FStart + FCount - 1) mod FMaxElements];
+    Result := FElements[Int64((FStart) + Int64(FCount) - 1) mod FMaxElements];
 end;
